Fix float/double pointer mismatches from the precision widening
3D selection stopped picking anything. VolumeCanIntersect{Box,Point} declare
local buffers for the float-only clipbox helpers, and 8ec36c5 widened them to
double -- so Matrix4GetFloats wrote 16 floats into a double[16] and
aabbToClipbox read doubles as floats. Those locals go back to float, which is
the narrowing boundary.

Three more of the same class, all silent because C only warns here:

- -scanFloat: into now-double fields (ROTSTEP angles, texture plane points,
  MLCad transforms) put 4 bytes in an 8-byte slot -- now -scanDouble:;
- &floatLocal passed to V3RayIntersects*'s double * out-params wrote 8 bytes
  into a 4-byte local -- those depth/tolerance locals are double now;
- LDrawRelatedPart's offset:(float *) received a double[3] -- widened to match.

So ROTSTEP rotations, texture planes and related-part offsets were broken in
that commit too.

Adds PrecisionBoundaryTests: the culling functions must accept an overlapping
screen box and reject a distant one, and a ROTSTEP line must survive scanning
plus the ZYX round-trip.

Found with `grep "incompatible pointer types"` over a clean build -- the build
succeeded and every test passed with these bugs present.

+++ trunk/Packages/LDrawCore/Sources/LDrawCore/LDrawDragHandle.m
@@ -181,8 +181,8 @@
 			hits:(NSMutableDictionary *)hits
 {
 	float   handleScale     = 0.0;
-	float   drawRadius      = 0.0;
-	float   intersectDepth  = 0;
+	double  drawRadius      = 0.0;
+	double  intersectDepth  = 0;
 	bool    intersects      = false;
 	
 	handleScale = 1.0 / scaleFactor;
